+++ src/aws/index.ts
@@ -1,6 +1,7 @@
 import { AppResources } from "./appResources";
 import { AppUser } from "./appUser";
 import { ECRRepository } from "./ecrRepository";
+import { GithubRunnerRole } from "./githubRunnerRole";
 import { GitlabRunnerAWSIAMUser } from "./gitlabRunnerAwsIamUser";
 import { S3Bucket } from "./s3Bucket";
 import { Secret } from "./secret";
@@ -9,6 +10,7 @@ export const aws = {
   AppResources,
   AppUser,
   ECRRepository,
+  GithubRunnerRole,
   GitlabRunnerAWSIAMUser,
   S3Bucket,
   Secret,